fixed frm return warehouse delete

+++ ServeAll.Core/Entities/ViewReturnWarehouse.cs
@@ -12,7 +12,7 @@ namespace ServeAll.Core.Entities
         public string product_code { get; set; }
         public string product_name { get; set; }
         public string return_number { get; set; }
-        public decimal return_quantity { get; set; }
+        public int return_quantity { get; set; }
         public string branch_details { get; set; }
         public string destination { get; set; }
         public DateTime return_date { get; set; }
